backend/requirements.txt
# FastAPI Core
# Using latest versions with Pydantic v2 support
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# AI & ML APIs
openai==1.14.0  # Updated for better Python 3.13 support
google-cloud-translate==3.15.0
google-cloud-language==2.13.0
google-generativeai==0.3.2
# anthropic==0.18.0  # Commented out - not used in current implementation

# AWS Services
boto3==1.34.34  # For Amazon Personalize, SES
aioboto3==12.3.0  # Async AWS

# Message Queue
aio-pika==9.3.1  # Async RabbitMQ

# Authentication & Security
# Note: Using pre-built cryptography version to avoid Rust compilation
cryptography==41.0.7  # Pre-built wheels available
python-jose==3.3.0  # Without [cryptography] extra to avoid version conflicts
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Task Scheduling
apscheduler==3.10.4

# Utilities
# Using Pydantic v2 with pre-built wheels
pydantic==2.7.0  # Has pre-built wheels for Python 3.13
pydantic-settings==2.2.0
email-validator==2.1.0
python-slugify==8.0.1

# Monitoring & Logging
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
